doc addition - changed neo4j version getting
