+++ core/src/main/java/apoc/cypher/CypherInitializer.java
@@ -2,8 +2,10 @@ package apoc.cypher;
 
 import apoc.ApocConfig;
 import apoc.util.Util;
+import apoc.version.Version;
 import org.apache.commons.configuration2.Configuration;
 import org.neo4j.common.DependencyResolver;
+import org.neo4j.configuration.Config;
 import org.neo4j.configuration.GraphDatabaseSettings;
 import org.neo4j.internal.helpers.collection.Iterators;
 import org.neo4j.kernel.api.procedure.GlobalProcedures;
@@ -15,6 +17,7 @@ import java.util.Collection;
 import java.util.Collections;
 
 import java.util.ConcurrentModificationException;
+import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
@@ -24,17 +27,20 @@ public class CypherInitializer implements AvailabilityListener {
     private final Log userLog;
     private final GlobalProcedures procs;
     private final DependencyResolver dependencyResolver;
+    private final String defaultDb;
 
     /**
      * indicates the status of the initializer, to be used for tests to ensure initializer operations are already done
      */
     private boolean finished = false;
+    private boolean checkedVersion = false;
 
     public CypherInitializer(GraphDatabaseAPI db, Log userLog) {
         this.db = db;
         this.userLog = userLog;
         this.dependencyResolver = db.getDependencyResolver();
         this.procs = dependencyResolver.resolveDependency(GlobalProcedures.class);
+        this.defaultDb = dependencyResolver.resolveDependency(Config.class).get(GraphDatabaseSettings.default_database);
     }
 
     public boolean isFinished() {
@@ -57,6 +63,23 @@ public class CypherInitializer implements AvailabilityListener {
                 if (!isSystemDatabase) {
                     awaitApocProceduresRegistered();
                 }
+
+                if (defaultDb.equals(db.databaseName())) {
+                    try {
+                        final String apocFullVersion = Version.class.getPackage().getImplementationVersion();
+                        final String apocVersion = getMajorMinVersion(apocFullVersion);
+                        final List<String> versions = db.executeTransactionally("CALL dbms.components", Collections.emptyMap(),
+                                r -> (List<String>) r.next().get("versions"));
+                        final boolean versionDifferent = versions.stream().noneMatch(kernelVersion -> getMajorMinVersion(kernelVersion).equals(apocVersion));
+                        if (versionDifferent) {
+                            userLog.warn("The apoc version (%s) and the Neo4j DBMS versions %s are incompatible. \n" +
+                                            "See the compatibility matrix in https://neo4j.com/labs/apoc/4.4/installation/ to see the correct version",
+                                    apocFullVersion, versions.toString());
+                        }
+                    } catch (Exception ignored) {
+                        // with embedded testdb, "call dbms.components" is not recognized here 
+                    }
+                }
                 Configuration config = dependencyResolver.resolveDependency(ApocConfig.class).getConfig();
 
                 for (String query : collectInitializers(isSystemDatabase, config)) {
@@ -73,6 +96,14 @@ public class CypherInitializer implements AvailabilityListener {
                 finished = true;
             }
         }).start();
+    }
+
+    private String getMajorMinVersion(String completeVersion) {
+        if (completeVersion == null) {
+            return null;
+        }
+        final String[] split = completeVersion.split("\\.");
+        return split[0] + "." + split[1];
     }
 
     private Collection<String> collectInitializers(boolean isSystemDatabase, Configuration config) {
